requirements.txt
colorama
gevent
IPy
loguru
lxml
ndjson
pwn
pwntools
pycryptodome
pyOpenSSL
requests
tzlocal
python-telegram-bot==20.7
Pillow
watchdog
